FEAT: Implement food deletion and enhance FoodListScreen UI
This commit introduces the ability to delete food items from the `FoodListScreen`. The UI has been improved with a new `FoodItemCard` component and a refactored `FoodListContent` for better organization. The "Edit" button now toggles an editing state, displaying delete icons on food items when active.

- Implemented food deletion functionality in `FoodsListViewmodel` and `DeleteFoodUseCase`.
- Created `FoodItemCard.kt` for displaying individual food items with an optional delete icon.
- Refactored `FoodListScreen.kt` by moving UI logic into `FoodListContent.kt`.
- Added an "editing" state to `FoodListScreen` that toggles the visibility of delete icons on `FoodItemCard`s.
- Updated `DashBoardScreen` FAB to navigate to "AddMeal" instead of "AddFood".
- Modified `DeleteFoodUseCase` to accept a food ID instead of a `FoodModel` object.
