+++ app/src/main/java/com/example/feedm/petsFeature/domain/objectTasks/food/useCase/DeleteFoodUseCase.kt
@@ -1,11 +1,14 @@
 package com.example.feedm.petsFeature.domain.objectTasks.food.useCase
 
-import com.example.feedm.petsFeature.domain.objectTasks.food.model.FoodModel
 import com.example.feedm.petsFeature.data.FoodRepository
 import javax.inject.Inject
 
-class DeleteFoodUseCase @Inject constructor(private val foodRepository: FoodRepository) {
-    suspend operator fun invoke(food: FoodModel) {
+class DeleteFoodUseCase @Inject constructor(
+    private val foodRepository: FoodRepository,
+    private val getFoodUseCase: GetFoodUseCase) {
+    suspend operator fun invoke(id: Int) {
+        val food = getFoodUseCase(id)
         foodRepository.deleteFood(food)
+
     }
 }
